fix: 按照审查路线图修复 P0/P1/P2 缺陷及部分 P3 项
P0: PillResistance @Id + ON CONFLICT upsert, SpiritEmotionTools ScopedValue
P1: @Transactional, @EnumValue, AuthenticatedAspect 错误码, Jackson 统一, auto-timestamp
P2: SpiritChatService 重构, BeastService 类型安全 VO, Command Handler 业务逻辑上移, Monster beans 迁移至 service/combat
P3: WeightedRandom throw, TypeUtils 委托, PlayerBuffRepository 改进, existsById→countByQuery, @EnableAspectJAutoProxy 移至主类

+++ src/main/java/top/stillmisty/xiantao/infrastructure/repository/FudiCellRepositoryImpl.java
@@ -62,6 +62,6 @@ public class FudiCellRepositoryImpl implements FudiCellRepository {
   @Override
   public int countByFudiId(Long fudiId) {
     QueryWrapper queryWrapper = QueryWrapper.create().where(FUDI_CELL.FUDI_ID.eq(fudiId));
-    return (int) fudiCellMapper.selectCountByQuery(queryWrapper);
+    return Math.toIntExact(fudiCellMapper.selectCountByQuery(queryWrapper));
   }
 }
